HolisticElectricityModel API Reference

HolisticElectricityModel.AbstractAgent — Type

abstract type AbstractAgent

Abstract type for agents.

Required interfaces:

get_id(agent::Agent)::String

getcurrentyear(agent::Agent, model_data::HEMData)::Tuple{Float64,Symbol}

solveagentproblem!( agents::AgentGroup, agentopts::AgentOptions, modeldata::HEMData, hemopts::HEMOptions, agentstore::AgentStore, )

saveresults( agent::AbstractAgent, agentopts::AgentOptions, hemopts::HEMOptions, exportfilepath::AbstractString, fileprefix::AbstractString,

)

HolisticElectricityModel.Agent — Type

Abstract type for all individual agents.

HolisticElectricityModel.AgentAndOptions — Type

struct AgentAndOptions{T<:AbstractAgent, U<:AgentOptions}

Struct to store an agent and its options.

agent::AbstractAgent

options::AgentOptions

HolisticElectricityModel.AgentGroup — Type

Abstract type for a group of individual agents.

HolisticElectricityModel.AgentOptionsStore — Type

Struct to store parsed agent options.

HolisticElectricityModel.AgentStore — Type

struct AgentStore

Struct to store a collection of agents and their options.

data::OrderedDict{DataType, OrderedDict{String, AgentAndOptions}}

This is passed to the agents to allow them to access other agents and their options.

HolisticElectricityModel.AgentStore — Method

AgentStore(
    agents_and_opts::Vector{AgentAndOptions}
) -> AgentStore

Create an AgentStore from a vector of AgentAndOptions.

HolisticElectricityModel.DERAggregatorOptions — Method

DERAggregatorOptions(
    attributes::MathOptInterface.OptimizerWithAttributes;
    incentive_curve,
    frac_viu_cost_savings_as_revenue
) -> DERAggregatorOptions

Construct DERAggregatorOptions with an MOI.OptimizerWithAttributes instance.

HolisticElectricityModel.Dimension — Type

Behaves like a Vector of Symbols with additional metadata fields.

HolisticElectricityModel.GreenDeveloperOptions — Method

GreenDeveloperOptions(
    attributes::MathOptInterface.OptimizerWithAttributes
) -> GreenDeveloperOptions

Construct GreenDeveloperOptions with an MOI.OptimizerWithAttributes instance.

HolisticElectricityModel.GurobiSolver — Type

Construct an Gurobi optimizer. Use default settings or pass an MOI.OptimizerWithAttributes instance.

If using the default settings, call import_gurobi() first. Gurobi must be installed.

Examples

julia> import_gurobi()
julia> hem_solver = GurobiSolver()

julia> import Gurobi
julia> import JuMP
julia> gurobi = JuMP.optimizer_with_attributes(
    () -> Gurobi.Optimizer(Gurobi.Env()),
    "MIPFocus" => 3,
)
julia> solver = GurobiSolver(gurobi)

HolisticElectricityModel.HEMData — Type

mutable struct HEMData

Struct to store the data required for the HEM model.

epsilon::ParamScalar: iteration tolerance

index_y::Dimension: Simulation years in a particular window (gets updated in solveequilibriumproblem!)

index_y_fix::Dimension: Represents the full simulation horizon (does not change). E.g., when we simulate years 2021-2030, "indexyfix" will be [2021, ..., 2030] if the planning window is 5-year for utility or IPPs, so the first indexy will be [2021, ..., 2025], after solving the first window, indexy will be updated to [2022, ..., 2026] etc.

index_d::Dimension: Representative day (from ReEDS)

index_t::Dimension: Representative hour index within each representative day

index_h::Dimension: Customer groups

index_z::Dimension: Zones (ReEDS BA modeled)

index_sector::Dimension: Customer high level groups

index_h_sector_map::DimensionSet: Map from customer group to sector

index_z_h_map::DimensionSet: Map from zone to customer group

omega::ParamArray: Number of days per representative day

year::ParamArray: Year

time::ParamArray: Time

year_start::ParamScalar: Simulation start year

delta_t::ParamScalar: Number of hours per representative hour

HolisticElectricityModel.HEMData — Method

HEMData(
    input_dir::String;
    year_start,
    delta_t,
    epsilon
) -> HEMData

Create a new HEMData instance.

HolisticElectricityModel.IPPOptions — Method

Construct IPPOptions with solvers defined as MOI.OptimizerWithAttributes instances or HEMSolver instances.

Review the functions that call get_new_jump_model for Dict key requirements.

Examples

julia> ipp_opts = IPPOptions(
    LagrangeDecomposition(),
    Dict(
        "Lagrange_Sub_Investment_Retirement_Cap" => JuMP.optimizer_with_attributes(
            Ipopt.Optimizer,
            "print_level" => 0,
        ),
        "Lagrange_Sub_Dispatch_Cap" => JuMP.optimizer_with_attributes(
            Xpress.Optimizer,
        ),
        "Lagrange_Feasible_Cap" => JuMP.optimizer_with_attributes(
            Xpress.Optimizer,
        )
    )
)

HolisticElectricityModel.IpoptSolver — Type

Construct an Ipopt optimizer. Use default settings or pass an MOI.OptimizerWithAttributes instance.

If using the default settings, call import_ipopt() first. Ipopt must be installed.

Examples

julia> import_ipopt()
julia> hem_solver = IpoptSolver()

julia> import Ipopt
julia> import JuMP
julia> ipopt = JuMP.optimizer_with_attributes(Ipopt.Optimizer, "tol" => 1e-6, "print_level" => 0)
julia> solver = IpoptSolver(ipopt)

HolisticElectricityModel.PVAdoptionModel — Method

PVAdoptionModel(
    Shape,
    MeanPayback,
    Bass_p,
    Bass_q
) -> HolisticElectricityModel.PVAdoptionModel

Constructs a PVAdoptionModel by computing Rate from the other parameters.

HolisticElectricityModel.ParamScalar — Type

Behaves like a number with additional metadata fields.

HolisticElectricityModel.UtilityOptions — Method

UtilityOptions(
    attributes::MathOptInterface.OptimizerWithAttributes
) -> UtilityOptions

Construct UtilityOptions with an MOI.OptimizerWithAttributes instance.

HolisticElectricityModel.XpressSolver — Type

Construct an Xpress optimizer. Use default settings or pass an MOI.OptimizerWithAttributes instance.

If using the default settings, call import_xpress() first. Xpress must be installed.

Examples

julia> import_xpress()
julia> hem_solver = XpressSolver()

julia> import Xpress
julia> import JuMP
julia> xpress = JuMP.optimizer_with_attributes(
    Xpress.Optimizer,
    "MIPRELSTOP" => 0.01,
    "OUTPUTLOG" => 1,
)
julia> solver = XpressSolver(xpress)

HolisticElectricityModel.configure_logging — Method

configure_logging(
;
    console_level,
    file_level,
    filename
) -> InfrastructureSystems.MultiLogger

Creates console and file loggers and sets the global_logger.

Note: Log messages may not be written to the file until flush() or close() is called on the returned logger.

Arguments

console_level = Logging.Error: level for console messages

file_level = Logging.Info: level for file messages

filename::Union{Nothing, AbstractString} = "hem.log": log file; pass nothing to disable file logging

Example

logger = configure_logging(console_level = Logging.Info)
@info "log message"
close(logger)

HolisticElectricityModel.get_agent — Method

get_agent(
    ::Type{T<:AbstractAgent},
    store::AgentStore
) -> AbstractAgent
get_agent(
    ::Type{T<:AbstractAgent},
    store::AgentStore,
    id
) -> AbstractAgent

Return the agent of the given type and ID from the store.

If there is only one agent of the given type then id is optional.

HolisticElectricityModel.get_agent_option — Method

get_agent_option(
    _::Type{T<:AbstractAgent},
    options::AgentOptionsStore
) -> AgentOptions

Returns the options for the agent type T from the options store. If no options are found for the agent type, an error is raised.

HolisticElectricityModel.get_bulk_system_agent — Method

get_bulk_system_agent(
    store::AgentStore,
    _::HEMOptions{VIU}
) -> AbstractAgent

Get the corresponding bulk system agent from the store based on the market structure within HEMOptions. If the market structure is VIU, it returns the Utility agent. If the market structure is WM, it returns the IPPGroup agent.

HolisticElectricityModel.get_current_year — Method

get_current_year(
    agent::AbstractAgent,
    model_data::HEMData
) -> Tuple{Any, Any}

Get the last year for which agent's data have been updated. Provided so other agents can access the right data for agent.

HolisticElectricityModel.get_one_to_many_dict — Method

get_one_to_many_dict(
    set::DimensionSet{N},
    i::Integer
) -> Dict{Any, Any}

Returns DimensionSet.set data as a Dict{Symbol,Vector} where the keys correspond to the ith dimension of the DimensionSet and the vector contains either Symbols (if N == 2) or Tuple{N-1,Symbol} (if N > 2).

HolisticElectricityModel.get_option — Method

get_option(
    ::Type{T<:AbstractAgent},
    store::AgentStore
) -> AgentOptions
get_option(
    ::Type{T<:AbstractAgent},
    store::AgentStore,
    id
) -> AgentOptions

Return the options for the agent of the given type and ID from the store. If there is only one agent of the given type then id is optional.

HolisticElectricityModel.get_prev_reg_year — Method

get_prev_reg_year(
    model_data::HEMData,
    w_iter::Integer
) -> Tuple{Any, Any}

Returns the current year and its Symbol representation from the model data, taking into account the window iteration.

HolisticElectricityModel.get_prev_two_reg_year — Method

get_prev_two_reg_year(
    model_data::HEMData,
    w_iter::Integer
) -> Tuple{Any, Any}

Returns the year two iterations before the current one and its Symbol representation, taking into account the window iteration. If w_iter is 2, it returns the previous year. If w_iter is 1, it returns the current year.

HolisticElectricityModel.get_reg_year — Method

get_reg_year(model_data::HEMData) -> Tuple{Any, Any}

Returns the current year and its Symbol representation from the model data.

HolisticElectricityModel.initialize_param — Method

initialize_param(
    name::AbstractString,
    index::Dimension;
    value,
    prose_name,
    description
) -> ParamArray{1}

Returns a ParamArray with all values set to value.

HolisticElectricityModel.initialize_param — Method

initialize_param(
    name::AbstractString,
    indices...;
    value,
    prose_name,
    description
) -> ParamArray{1}

Return a ParamArray with all values set to value, and indices formed from Iterators.product(indices...).

HolisticElectricityModel.ipp_cap_lower — Method

ipp_cap_lower(
    ipp,
    ipp_opts,
    model_data,
    delta_t,
    window_length,
    customers,
    der_aggregator,
    green_developer,
    solver
) -> Any

Lower level optimization results are used to set variable bounds for McCormick-envelope Relaxation

HolisticElectricityModel.ipp_cap_lower_dual — Method

ipp_cap_lower_dual(
    ipp,
    ipp_opts,
    model_data,
    delta_t,
    window_length,
    customers,
    der_aggregator,
    green_developer,
    solver
) -> Any

Lower level dual optimization results are used to set variable bounds for McCormick-envelope Relaxation

HolisticElectricityModel.iter_agents_and_options — Method

iter_agents_and_options(
    store::AgentStore
) -> Base.Iterators.Flatten{Base.Generator{Base.ValueIterator{OrderedDict{DataType, OrderedDict{String, AgentAndOptions}}}, HolisticElectricityModel.var"#117#118"}}

Iterate over all agents and their options in the store and return a Tuple of the agent and its options.

HolisticElectricityModel.make_keyed_array — Method

make_keyed_array(indices...; fill_nan) -> Any

Return an uninitialized KeyedArray from any number of Dimension values.

HolisticElectricityModel.read_keyed_array — Method

read_keyed_array(
    filename::AbstractString,
    num_dims::Int64
) -> Any

Return a KeyedArray from an N-dimensional array flattened in a CSV file.

If there is one dimension then the file must have a single row with dimension names. If there is more than one dimension then it must conform to the following format:

The header row consists of dimension names for the first N-1 dimensions followed by the last dimension's element ids pivoted out to form data column headers.

The data rows contain dimension element ids in the first N-1 columns followed by parameter values in the remaining columns. Each value maps to the dimension element ids listed in the row's first N-1 columns plus the dimension element id found in that value's column.

3-dimension example:

d1variablename,d2variablename,d3name1,d3name2,d3name3 d1name1,d2name1,1.0,1.0,1.0 d1name2,d2_name2,1.0,1.0,1.0

HolisticElectricityModel.read_param — Method

read_param(
    name::AbstractString,
    dirpath::AbstractString,
    filename::AbstractString,
    column_index::Dimension,
    row_indices::Vector{Dimension};
    prose_name,
    description
) -> ParamArray

Reads parameter data from the csv file at dirpath/filename. Assumes the data has a single header row, and that the columnindex elements are the right-most column names. All other data dimension values are listed in the first length(rowindices) columns of the file, corresonding one-to-one and in the same order as the row_indices.

Returns the data loaded into a ParamAxisArray with dimensions (rowindices..., columnindex).

HolisticElectricityModel.read_param — Method

read_param(
    name::AbstractString,
    dirpath::AbstractString,
    filename::AbstractString,
    index::Dimension;
    prose_name,
    description
) -> ParamArray{1}

Reads parameter data from the csv file at dirpath/filename. Assumes the data has a single header row, and that the index elements comprise the column names.

Returns the data loaded into a ParamArray.

HolisticElectricityModel.read_set — Method

read_set(
    name::AbstractString,
    dirpath::AbstractString,
    filename::AbstractString,
    dims::Vector{Dimension};
    prose_name,
    description
) -> DimensionSet

Reads DimensionSet data from dirpath/filename.

Returns a DimensionSet

HolisticElectricityModel.read_set — Method

read_set(
    dirpath::AbstractString,
    filename::AbstractString,
    name::AbstractString;
    prose_name,
    description
) -> Dimension

Reads Dimension data from dirpath/filename.

HolisticElectricityModel.run_hem — Method

run_hem(
    input_dir::AbstractString,
    options::HEMOptions;
    agent_options,
    max_iterations,
    window_length,
    force,
    jump_model,
    output_dir,
    delta_t,
    year_start
)

Solve the problem with the given inputs.

Arguments

input_dir::AbstractString: Directory containing input data. Outputs will be recorded in a subdirectory.

options::HEMOptions:

ipp_options::IPPOptions: Independent Power Producer options. Defaults to NullAgentOptions.

regulator_options::RegulatorOptions: Regulator options. Defaults to NullAgentOptions.

utility_options::RegulatorOptions: Regulator options. Defaults to NullAgentOptions.

max_iterations::Int: Max number of iterations to attempt a solution. Defaults to 100.

window_length::Int:

force::Bool: If true, overwrite results if a directory already exists.

jump_model::Any: Jump model to use for the simulation.

output_dir::Union{<:AbstractString, Nothing}: Directory to save results. If nothing, it will be set to input_dir with a prefix based on the options and agents.

delta_t::Int: Number of hours per representative hour. Defaults to 4.

year_start::Int: The base year for the simulation. Defaults to 2020.

HolisticElectricityModel.save_results — Method

save_results(
    agent::AbstractAgent,
    agent_opts::AgentOptions,
    hem_opts::HEMOptions,
    export_file_path::AbstractString
) -> DataFrames.DataFrame

This method needs to be implemented by each agent type to save its results.

HolisticElectricityModel.set_log_level — Method

set_log_level(level::Base.CoreLogging.LogLevel)

Sets the log level to, e.g., Logging.Debug, Logging.Warn. Currently only implemented for ConsoleLogger.

HolisticElectricityModel.solve_equilibrium_problem! — Method

solve_equilibrium_problem!(
    hem_opts::HEMOptions,
    model_data::HEMData,
    agents_and_opts::Vector{AgentAndOptions},
    export_file_path::AbstractString,
    max_iter::Int64,
    window_length::Int64,
    jump_model
)

Main function to run the HEM model. It loops over the simulation years and calls the solve_agent_problem! method and the save_results method for each agent. Arguments:

input_dir::AbstractString: Directory containing input data. Outputs will be recorded in a subdirectory.

model_data::HEMData: Data required for the model.

hem_opts::HEMOptions: Options for the model.

agents_and_opts::Vector{AgentAndOptions}: Vector of agents and their options.

export_file_path::AbstractString: Path to export the results.

max_iter::Int64: Maximum number of iterations to solve the problem. Set in run_hem.

window_length::Int64: Length of the window for the simulation. Set in run_hem.

jump_model::Any: Jump model to use for the simulation. Iniitialized in run_hem.

HolisticElectricityModel.solve_green_developer_problem — Function

solve_green_developer_problem(
    green_developer::GreenDeveloper,
    green_developer_opts::GreenDeveloperOptions,
    model_data::HEMData,
    hem_opts::HEMOptions,
    agent_store::AgentStore,
    w_iter::Int64
) -> Any
solve_green_developer_problem(
    green_developer::GreenDeveloper,
    green_developer_opts::GreenDeveloperOptions,
    model_data::HEMData,
    hem_opts::HEMOptions,
    agent_store::AgentStore,
    w_iter::Int64,
    p::Union{Nothing, Symbol}
) -> Any

Create the green developer optimization problem depending on the market structure
and solve for the optimal green tech buildout. Calculate the ppa price from the optimization results.

HolisticElectricityModel.update_cumulative! — Method

update_cumulative!(
    model_data::HEMData,
    agent::AbstractAgent
)

Default no-op method for updating cumulative parameters in AbstractAgents after each model year.

HolisticElectricityModel.update_cumulative! — Method

update_cumulative!(model_data::HEMData, ipp::IPPGroup)

Update IPPGroup cumulative parameters

HolisticElectricityModel.update_cumulative! — Method

update_cumulative!(model_data::HEMData, utility::Utility)

Update Utility cumulative parameters

HolisticElectricityModel.update_total_capacity! — Method

update_total_capacity!(
    total_capacity_my::ParamArray,
    x_DG_new::ParamArray,
    model_data::HEMData,
    reg_year::Integer,
    z::Symbol,
    h::Symbol,
    m::Symbol
)

Calculates cumulative capacity for this year and future years if this technology was a build option.